+++ vignettes-raw/1-spark-cleanup.R
@@ -1,7 +1,5 @@
 unlink("derby.log")
-unlink("logs"
-       , recursive = TRUE, force = TRUE)
+unlink("logs", recursive = TRUE, force = TRUE)
 
 unlink(file.path("vignettes-raw", "derby.log"))
-unlink(file.path("vignettes-raw", "logs")
-       , recursive = TRUE, force = TRUE)
+unlink(file.path("vignettes-raw", "logs"), recursive = TRUE, force = TRUE)