+++ NEWS.md
@@ -1,3 +1,5 @@
+# surveytable (development version)
+
 # surveytable 0.9.7
 
 * Bugfixes.